Refactor position calculation in Player's move method
Remove unnecessary casting to long during position calculation in the move method. This simplifies the code and reduces potential type conversion issues.

+++ src/main/java/edu/ntnu/idi/bidata/Player.java
@@ -45,7 +45,7 @@ public class Player {
    * @param steps the number of steps to move forward
    */
   public void move(int steps) {
-    int newPosition = Math.clamp((int) ((long) currentTile.getPosition() + steps), 0, board.getNumberOfTiles() - 1);
+    int newPosition = Math.clamp((currentTile.getPosition() + steps), 0, board.getNumberOfTiles() - 1);
     Tiles newTile = board.getTile(newPosition);
     placeOnTile(newTile);
     newTile.landPlayer(this);
